restaurantActivity 메뉴 레이아웃 구성 완료 (리뷰부분만 추후 추가)

+++ app/src/main/java/com/example/risingtest/src/main/restaurant/models/RestInfoResult.kt
@@ -7,7 +7,7 @@ data class RestInfoResult(
     @SerializedName("restName") val restName : String,
     @SerializedName("star") val star : Double,
     @SerializedName("reviewCount") val reviewCount : Int,
-    @SerializedName("distance_KM") val distance_KM : Int,
+    @SerializedName("distance_KM") val distance_KM : Double,
     @SerializedName("Cheetah") val Cheetah : String,
     @SerializedName("deliveryFee") val deliveryFee : String,
     @SerializedName("minimumAmount") val minimumAmount : String
